Moving edit and save files to components dir
